Move BalanceSelector item into its own component and the title inside. Moved BalanceSelector to its own folder.
