+++ src/containers/Accounts/AccountHeader/BalanceSelector/balance-selector.scss
@@ -1,4 +1,4 @@
-@import '../../shared/css/variables';
+@import 'src/containers/shared/css/variables';
 
 .balance-selector {
   position: relative;
